requirements.txt
Django==4.2.7
graphene-django==3.1.5
django-graphql-jwt==0.4.0
psycopg2-binary
channels==3.0.5
channels-redis==3.4.1
django-filter==23.3
django-cors-headers==4.3.0
factory-boy==3.3.0
pytest-django==4.5.2
python-dotenv==1.0.0
sentry-sdk==1.38.0
PyJWT==2.8.0
redis==5.0.1
daphne==3.0.2
graphene-subscriptions==1.0.2
